validate networks after .. still wip

+++ driver/driver.go
@@ -293,7 +293,7 @@ func (d *driver) CreatePersistentNetworks() error {
 	}
 
 	for _, n := range nwList {
-		if IsNetworkIdValid(n.NetworkID) == false {
+		if false && IsNetworkIdValid(n.NetworkID) == false {
 			nwDir := filepath.Join(persistConfigPath, n.NetworkID)
 			os.RemoveAll(nwDir)
 			log.Println("Skipping and deleting stale network: ", n.NetworkID)
@@ -313,6 +313,25 @@ func (d *driver) CreatePersistentNetworks() error {
 	}
 	return nil
 }
+
+func (d *driver) ValidatePersistentNetworks() error {
+	nwList, err := Read_Past_Config(persistConfigPath)
+	if err != nil {
+		return err
+	}
+
+	for _, n := range nwList {
+		if IsNetworkIdValid(n.NetworkID) == false {
+			nwDir := filepath.Join(persistConfigPath, n.NetworkID)
+			os.RemoveAll(nwDir)
+			log.Println("Skipping and deleting stale network: ", n.NetworkID)
+			continue
+		}
+	}
+	return nil
+}
+
+
 
 func StartDriver() (*driver, error) {
